Refactor the PyodideRuntimeAgent to be a subclass of RuntimeAgent
