+++ packages/lib/mod.ts
@@ -34,3 +34,5 @@ export type { LoggerConfig } from "./src/logging.ts";
 
 // Media types and utilities for rich content handling
 export * from "./src/media/mod.ts";
+
+export { runner } from "./src/runtime-runner.ts";
